requirements.txt
chainlit
langchain
langchain-core
langgraph
plotly
langchain-ollama
langchain-mistralai
langchain-community
langgraph_supervisor
langchain-aws
psycopg[binary]
pandas
pydantic
python-dotenv
